requirements.txt
alembic==1.9.3
anyio==3.6.2
beautifulsoup4==4.11.2
black==23.1.0
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==39.0.1
ecdsa==0.18.0
fastapi==0.90.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
packaging==23.0
passlib==1.7.4
pathspec==0.11.0
platformdirs==3.0.0
psycopg==3.1.8
psycopg-binary==3.1.8
psycopg-pool==3.1.6
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.4
python-jose==3.3.0
python-multipart==0.0.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.2
starlette==0.23.0
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
uvicorn==0.20.0
